Реорганизует структуру компонентов
